refactor(ui): Conditionally switch out middleware depending on auth config
